+++ src/test/java/com/pepero/jcb/api/PGNDatabaseTest.java
@@ -7,13 +7,28 @@ import java.nio.file.Path;
 
 public class PGNDatabaseTest {
     public static void main(String[] args) {
-        try(PGNDatabaseIterator pgnDatabaseIterator = new PGNDatabaseIterator(Path.of("pgn/GM_Games.pgn"))) {
+        try(PGNDatabaseIterator pgnDatabaseIterator = new PGNDatabaseIterator(Path.of("D:\\PGN\\lichess_rated_2016-07.pgn"))) {
+            int games = 0;
+
+            long startTime = System.currentTimeMillis();
+
+            long time = System.currentTimeMillis();
+
             while (pgnDatabaseIterator.hasNext()) {
+                games++;
                 PGNGameStub gameStub = pgnDatabaseIterator.next();
-                System.out.println(ChessGame.fromPGN(gameStub.rawPGN()).getPGN());
-                System.out.println();
-                System.out.println("Reconverted PGN END");
+                gameStub.toChessGame();
+                if(games % 10000 == 0) {
+                    long stopTime = System.currentTimeMillis();
+                    System.out.println(games + " games searched...");
+                    System.out.println(((stopTime - time) / 1000.) + " seconds");
+                }
             }
+
+            long stopTime = System.currentTimeMillis();
+
+            System.out.println(games + " games searched.");
+            System.out.println("Time elapsed : " + ((stopTime - startTime) / 1000.) + " seconds");
         }
     }
 }